Updated the MainGUI to have more functionality
MainGUI now opens up the JFrame for the UserInputGUI when called. Adjusted placement of the combo boxes and buttons for WorkoutsGUI and WorkoutStartGUI. 
Allowed for UserInputGUI to create a User object when filled out by the user. Figured out how to have combo boxes and images one the same scene.